+++ src/Blitter.Blocks/PlayField2D.cs
@@ -1,0 +1,9 @@
+namespace Blitter.Blocks;
+
+public class PlayField2D: Container2D
+{
+    public PlayField2D(IEnumerable<Sprite2D> sprites) 
+        : base(sprites)
+    {
+    }
+}